Add configuration for codecov to stop spamming comment updates
Forwarding comment edit notifications into our internal chat means
that we need to tune codecov down to reduce notification spam. It
also means it would not comment if there is an issue with any of the
cells.

As a downside, this number needs to be adjusted every time the number
of cells in travis and appveyor changes.